// pro/src/apiClient/v1/models/CollectiveAdditionalFeeType.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum CollectiveAdditionalFeeType {
  ACCOMMODATION = 'ACCOMMODATION',
  TRAVEL = 'TRAVEL',
  MEAL = 'MEAL',
  CONSUMABLE_ITEMS = 'CONSUMABLE_ITEMS',
  COPYRIGHT = 'COPYRIGHT',
  APPLICATION_FEE = 'APPLICATION_FEE',
  OTHER = 'OTHER',
}